bridge/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.3
datetime==4.4
flask==2.1.2
idna==3.3; python_version >= '3'
importlib-metadata==4.11.3; python_version < '3.10'
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
urllib3==1.26.9
werkzeug==2.1.2
zipp==3.8.0
zope.interface==5.4.0
